Move classes around to fix classycle
