+++ subprojects/base-services-groovy/src/test/groovy/org/gradle/util/ConfigureUtilTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -190,11 +190,10 @@ class ConfigureUtilTest extends Specification {
         bean.prop == "foo"
     }
 
-}
-
-class Bean {
-    String prop
-    def method(String value) {
-        prop = value
+    static class Bean {
+        String prop
+        def method(String value) {
+            prop = value
+        }
     }
 }
