+++ subprojects/base-services/src/test/groovy/org/gradle/util/CollectionUtilsTest.groovy
@@ -17,13 +17,33 @@ package org.gradle.util
 
 import org.gradle.api.Action
 import org.gradle.api.Transformer
-import org.gradle.api.internal.ClosureBackedAction
 import org.gradle.api.specs.Spec
 import org.gradle.api.specs.Specs
 import spock.lang.Specification
 import spock.lang.Unroll
 
-import static org.gradle.util.CollectionUtils.*
+import static org.gradle.util.CollectionUtils.addAll
+import static org.gradle.util.CollectionUtils.collect
+import static org.gradle.util.CollectionUtils.collectMap
+import static org.gradle.util.CollectionUtils.collectMapValues
+import static org.gradle.util.CollectionUtils.compact
+import static org.gradle.util.CollectionUtils.diffSetsBy
+import static org.gradle.util.CollectionUtils.every
+import static org.gradle.util.CollectionUtils.filter
+import static org.gradle.util.CollectionUtils.flattenCollections
+import static org.gradle.util.CollectionUtils.groupBy
+import static org.gradle.util.CollectionUtils.inject
+import static org.gradle.util.CollectionUtils.intersection
+import static org.gradle.util.CollectionUtils.join
+import static org.gradle.util.CollectionUtils.nonEmptyOrNull
+import static org.gradle.util.CollectionUtils.partition
+import static org.gradle.util.CollectionUtils.replace
+import static org.gradle.util.CollectionUtils.sort
+import static org.gradle.util.CollectionUtils.stringize
+import static org.gradle.util.CollectionUtils.toList
+import static org.gradle.util.CollectionUtils.toSet
+import static org.gradle.util.CollectionUtils.toStringList
+import static org.gradle.util.CollectionUtils.unpack
 
 @Unroll
 class CollectionUtilsTest extends Specification {
