+++ subprojects/core/src/integTest/groovy/org/gradle/internal/scopeids/ContinuousScopeIdsIntegrationTest.groovy
@@ -18,10 +18,10 @@
 package org.gradle.internal.scopeids
 
 import org.gradle.api.Action
-import org.gradle.api.internal.ClosureBackedAction
 import org.gradle.integtests.fixtures.AbstractContinuousIntegrationTest
 import org.gradle.integtests.fixtures.ScopeIdsFixture
 import org.gradle.integtests.fixtures.executer.GradleExecuter
+import org.gradle.util.ClosureBackedAction
 import org.junit.Rule
 
 class ContinuousScopeIdsIntegrationTest extends AbstractContinuousIntegrationTest {
