+++ subprojects/core/src/main/java/org/gradle/api/internal/file/copy/DelegatingCopySpecInternal.java
@@ -24,8 +24,8 @@ import org.gradle.api.file.CopySpec;
 import org.gradle.api.file.DuplicatesStrategy;
 import org.gradle.api.file.FileCopyDetails;
 import org.gradle.api.file.FileTreeElement;
-import org.gradle.api.internal.ClosureBackedAction;
 import org.gradle.api.specs.Spec;
+import org.gradle.util.ClosureBackedAction;
 
 import javax.annotation.Nullable;
 import java.io.FilterReader;
