+++ subprojects/core/src/test/groovy/org/gradle/api/internal/PropertiesTransformerTest.groovy
@@ -16,6 +16,7 @@
 package org.gradle.api.internal
 
 import org.gradle.api.Action
+import org.gradle.util.ClosureBackedAction
 import spock.lang.Specification
 
 class PropertiesTransformerTest extends Specification {
