+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/properties/bean/NestedRuntimeBeanNodeTest.groovy
@@ -17,7 +17,7 @@
 package org.gradle.api.internal.tasks.properties.bean
 
 import org.gradle.api.Action
-import org.gradle.api.internal.ClosureBackedAction
+import org.gradle.util.ClosureBackedAction
 import org.gradle.util.ConfigureUtil
 import spock.lang.Specification
 import spock.lang.Unroll
