+++ subprojects/ide/src/main/java/org/gradle/plugins/ide/eclipse/model/EclipseProject.java
@@ -23,9 +23,9 @@ import groovy.lang.DelegatesTo;
 import org.gradle.api.Action;
 import org.gradle.api.Incubating;
 import org.gradle.api.InvalidUserDataException;
-import org.gradle.api.internal.ClosureBackedAction;
 import org.gradle.plugins.ide.api.XmlFileContentMerger;
 import org.gradle.plugins.ide.eclipse.model.internal.DefaultResourceFilter;
+import org.gradle.util.ClosureBackedAction;
 
 import javax.inject.Inject;
 import java.util.Arrays;
