+++ subprojects/maven/src/main/java/org/gradle/api/publication/maven/internal/deployer/DefaultGroovyMavenDeployer.java
@@ -20,11 +20,11 @@ import groovy.lang.Closure;
 import org.apache.maven.artifact.ant.RemoteRepository;
 import org.gradle.api.artifacts.maven.GroovyMavenDeployer;
 import org.gradle.api.artifacts.maven.PomFilterContainer;
-import org.gradle.api.internal.ClosureBackedAction;
 import org.gradle.api.internal.artifacts.mvnsettings.LocalMavenRepositoryLocator;
 import org.gradle.api.internal.artifacts.mvnsettings.MavenSettingsProvider;
 import org.gradle.api.publication.maven.internal.ArtifactPomContainer;
 import org.gradle.internal.logging.LoggingManagerInternal;
+import org.gradle.util.ClosureBackedAction;
 import org.gradle.util.ConfigureUtil;
 
 import java.util.Map;
